physical-ai-challenge-2026/macos_pipeline/requirements.txt
ultralytics
onnx
onnxruntime
open3d
opencv-python
numpy
matplotlib
pillow
scipy
mujoco
